+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgdown 1.1.0.9000
 
+* Better conversion for `\tabular{}` tags containing empty cells (#780).
+
 * All third party resources are now fetched from a single CDN and are associated
   with a SRI hash (@bisaloo, #893).